requirements.txt
Flask==3.0.0
tensorflow-cpu==2.16.1
Pillow==10.2.0
numpy==1.26.3